requirements.txt
base45==0.4.1
cbor2==5.4.1
Pillow==8.3.1
pyzbar==0.1.8
